+++ cli.js
@@ -43,7 +43,8 @@ var cli = meow({
         'basic-text',
         'show-time',
         'deflate-woff',
-        'css-glyph'
+        'css-glyph',
+        'version'
     ],
     'string': [
         'text',
@@ -54,9 +55,16 @@ var cli = meow({
         b: 'basic-text',
         d: 'deflate-woff',
         T: 'show-time',
-        h: 'help'
+        h: 'help',
+        v: 'version'
     }
 });
+
+// version
+if (cli.flags.version) {
+    console.log(require('./package.json').version);
+    process.exit(0);
+}
 
 function isFile(path) {
     if (/^[^\s]+\.\w*$/.test(path)) {
@@ -87,6 +95,7 @@ function run(src, dest) {
 
     var fontmin = new Fontmin()
         .src(src)
+        .use(Fontmin.otf2ttf(pluginOpts))
         .use(Fontmin.glyph(pluginOpts))
         .use(Fontmin.ttf2eot(pluginOpts))
         .use(Fontmin.ttf2svg(pluginOpts))