+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## 0.9.1 (2015-12-22)
+- subset at fonteditor-core
+- performance optimization thx @akira-cn
+
+## 0.9.0-alpha-3 (2015-07-30)
+
+- fix css plugin glyf bug
+- up dependence fonteditor-core
+
+## 0.9.0-alpha-2 (2015-05-20)
+
+- css plugin add option local
+
+## 0.9.0-alpha-1 (2015-05-06)
+
+- add otf2ttf plugin
+
 ## 0.8.1 (2015-05-05)
 
 - svgs2ttf plugin add option fontName